moved audio pipeline into its own sub package
